completed feature to generate new rocket parameters, and save all data to output file
output file is named outputTest2.csv inside NN Training. When you run CDTester.py, is will generate the output file with as many rows as you run from the D2 Hypercube xlsx file. Currently the first row is set to 2, and last row is set to 3, but really the last row should be set to 12000. CORRECTIONS THAT NEED TO BE MADE: Rocket dict wind speed parameter is fixed which is wrong, and rockets are really janky.